feat: move delay fn in a root shared folder
