+++ renderer/src/common/lib/polling.ts
@@ -3,8 +3,7 @@
  */
 
 import type { WorkloadsWorkload } from '../api/generated/types.gen'
-
-const delay = (ms: number) => new Promise((resolve) => setTimeout(resolve, ms))
+import { delay } from '../../../../utils/delay'
 
 // Types
 interface PollingConfig {
